+++ LibCarla/source/carla/geom/Transform.h
@@ -21,25 +21,35 @@ namespace geom {
   class Transform {
   public:
 
+    // =========================================================================
+    // -- Public data members --------------------------------------------------
+    // =========================================================================
+
+    Location location;
+
+    Rotation rotation;
+
+    MSGPACK_DEFINE_ARRAY(location, rotation);
+
+    // =========================================================================
+    // -- Constructors ---------------------------------------------------------
+    // =========================================================================
+
     Transform() = default;
 
     Transform(const Location &in_location, const Rotation &in_rotation)
       : location(in_location),
         rotation(in_rotation) {}
 
-    Location location;
-    Rotation rotation;
+    // =========================================================================
+    // -- Other methods --------------------------------------------------------
+    // =========================================================================
 
-    bool operator==(const Transform &rhs) const {
-      return (location == rhs.location) && (rotation == rhs.rotation);
-    }
-
-    bool operator!=(const Transform &rhs) const {
-      return !(*this == rhs);
+    Vector3D GetForwardVector() const {
+      return rotation.GetForwardVector();
     }
 
     void TransformPoint(Vector3D &in_point) const {
-
       // Rotate
       double cy = cos(Math::to_radians(rotation.yaw));
       double sy = sin(Math::to_radians(rotation.yaw));
@@ -66,6 +76,22 @@ namespace geom {
       in_point = out_point;
     }
 
+    // =========================================================================
+    // -- Comparison operators -------------------------------------------------
+    // =========================================================================
+
+    bool operator==(const Transform &rhs) const {
+      return (location == rhs.location) && (rotation == rhs.rotation);
+    }
+
+    bool operator!=(const Transform &rhs) const {
+      return !(*this == rhs);
+    }
+
+    // =========================================================================
+    // -- Conversions to UE4 types ---------------------------------------------
+    // =========================================================================
+
 #ifdef LIBCARLA_INCLUDED_FROM_UE4
 
     Transform(const FTransform &transform)
@@ -77,8 +103,6 @@ namespace geom {
     }
 
 #endif // LIBCARLA_INCLUDED_FROM_UE4
-
-    MSGPACK_DEFINE_ARRAY(location, rotation);
   };
 
 } // namespace geom