+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 ## Latest changes
 
+  * Upgraded to Unreal Engine 4.21
+  * Upgraded Boost to 1.69.0
   * Added point transformation functionality for LibCarla and PythonAPI
   * Added "sensor_tick" attribute to sensors (cameras and lidars) to specify the capture rate in seconds
+  * Added "get_forward_vector()" to rotation and transform, retrieves the unit vector on the rotation's X-axis
   * Added support for Deepin in PythonAPI's setup.py
+  * Added support for spawning and controlling walkers (pedestrians)
+  * Renamed vehicle.get_vehicle_control() to vehicle.get_control() to be consistent with walkers
+  * Added support for gnss_sensor
 
 ## CARLA 0.9.2